requirements.txt
tensorflow==2.13.1
fastapi
matplotlib==3.8.2
python-dotenv==1.0.0
requests==2.31.0
uvicorn
onnxruntime
python-jose
passlib
torchvision
python-multipart
bcrypt
opencv-python
dlib
scipy
imutils